Move greeting message, cards, recipe tags, and hide functions from scripts.js to domUpdates.js
